+++ src/euklid/vector/vector.cpp
@@ -1,44 +1,44 @@
 #include "vector.hpp"
 
-template<unsigned int dimensions, typename T>
+template<size_t dimensions, typename T>
 Vector<dimensions, T>::Vector(const T& in) {
-    for(unsigned int i=0; i<dimensions; i++){
+    for(size_t i=0; i<dimensions; i++){
         this->set_item(i, in.get_item(i));
     }
 }
 
 
-template<unsigned int dimensions, typename T>
+template<size_t dimensions, typename T>
 Vector<dimensions, T>::Vector() {
-    for(unsigned int i=0; i<dimensions; i++){
+    for(size_t i=0; i<dimensions; i++){
         this->set_item(i, 0);
     }
 }
 
-template<unsigned int dimensions, typename T>
+template<size_t dimensions, typename T>
 Vector<dimensions, T>::Vector(double value) {
-    for(unsigned int i=0; i<dimensions; i++){
+    for(size_t i=0; i<dimensions; i++){
         this->set_item(i, value);
     }
 }
 
 /*
-template<unsigned int dimensions, typename T>
+template<size_t dimensions, typename T>
 double Vector<dimensions, T>::dot(const Vector<dimensions, T>& v2) {
     double result = 0;
 
-    for (unsigned int i=0; i<dimensions; i++) {
+    for (size_t i=0; i<dimensions; i++) {
         result += this->get_item(i) * v2.get_item(i);
     }
 
     return result;
 }*/
 
-template<unsigned int dimensions, typename T>
+template<size_t dimensions, typename T>
 double Vector<dimensions, T>::dot(const T& v2) {
     double result = 0;
 
-    for (unsigned int i=0; i<dimensions; i++) {
+    for (size_t i=0; i<dimensions; i++) {
         result += this->get_item(i) * v2.get_item(i);
     }
 
@@ -46,12 +46,12 @@ double Vector<dimensions, T>::dot(const T& v2) {
 }
 
 
-template<unsigned int dimensions, typename T>
+template<size_t dimensions, typename T>
 double Vector<dimensions, T>::length(){
     return sqrt(this->dot(static_cast<T>(*this)));
 }
 
-template<unsigned int dimensions, typename T>
+template<size_t dimensions, typename T>
 double Vector<dimensions, T>::get_item(int n) const {
     if (n < 0 || n >= (int)dimensions) {
         throw "Invalid dimension";
@@ -61,7 +61,7 @@ double Vector<dimensions, T>::get_item(int n) const {
 }
 
 
-template<unsigned int dimensions, typename T>
+template<size_t dimensions, typename T>
 void Vector<dimensions, T>::set_item(int n, double value) {
     if (n < 0 || n >= (int)dimensions) {
         throw "Invalid dimension";
@@ -70,7 +70,7 @@ void Vector<dimensions, T>::set_item(int n, double value) {
     this->coordinates[n] = value;
 }
 
-template<unsigned int dimensions, typename T>
+template<size_t dimensions, typename T>
 double& Vector<dimensions, T>::operator[](int n) {
     if (n < 0 || n >= (int)dimensions) {
         throw "Invalid dimension";
@@ -80,38 +80,38 @@ double& Vector<dimensions, T>::operator[](int n) {
 }
 
 
-template<unsigned int dimensions, typename T>
+template<size_t dimensions, typename T>
 double Vector<dimensions, T>::operator[](int n) const {
     return this->get_item(n);
 }
 
-template<unsigned int dimensions, typename T>
+template<size_t dimensions, typename T>
 T Vector<dimensions, T>::operator+(const T& v2) const {
     T result;
 
-    for (unsigned int i=0; i<dimensions; i++){
+    for (size_t i=0; i<dimensions; i++){
         result.set_item(i, this->get_item(i) + v2.get_item(i));
     }
 
     return result;
 }
 
-template<unsigned int dimensions, typename T>
+template<size_t dimensions, typename T>
 T Vector<dimensions, T>::operator-(const T& v2) const {
     T result;
 
-    for(unsigned int i=0; i<dimensions; i++) {
+    for(size_t i=0; i<dimensions; i++) {
         result.set_item(i, this->get_item(i) - v2.get_item(i));
     }
 
     return result;
 }
 
-template<unsigned int dimensions, typename T>
+template<size_t dimensions, typename T>
 T Vector<dimensions, T>::operator*(const T& v2) const {
     T result;
 
-    for (unsigned int i=0; i<dimensions; i++){
+    for (size_t i=0; i<dimensions; i++){
         result.set_item(i, this->get_item(i) * v2.get_item(i));
     }
 
@@ -119,32 +119,32 @@ T Vector<dimensions, T>::operator*(const T& v2) const {
 }
 
 
-template<unsigned int dimensions, typename T>
+template<size_t dimensions, typename T>
 T Vector<dimensions, T>::operator* (const double& factor) const {
     T result;
 
-    for (unsigned int i=0; i<dimensions; i++) {
+    for (size_t i=0; i<dimensions; i++) {
         result.set_item(i, this->get_item(i)*factor);
     }
 
     return result;
 }
 
-template<unsigned int dimensions, typename T>
+template<size_t dimensions, typename T>
 double Vector<dimensions, T>::distance(const T& v2) {
     return ((*this) - v2).length();
 }
 
-template<unsigned int dimensions, typename T>
+template<size_t dimensions, typename T>
 T Vector<dimensions, T>::copy() {
     return T(*this);
 }
 
-template<unsigned int dimensions, typename T>
+template<size_t dimensions, typename T>
 void Vector<dimensions, T>::normalize() {
     double len = this->length();
 
-    for (unsigned int i=0; i<dimensions; i++) {
+    for (size_t i=0; i<dimensions; i++) {
         this->set_item(i, this->get_item(i)/len);
     }
 }
@@ -179,12 +179,12 @@ Vector2D::Vector2D() {
 
 
 Vector2D::Vector2D(const Vector<2, Vector2D>& in) {
-    for(unsigned int i=0; i<Vector2D::dimension; i++){
+    for(size_t i=0; i<Vector2D::dimension; i++){
         this->set_item(i, in.get_item(i));
     }
 }
 Vector3D::Vector3D(const Vector<3, Vector3D>& in) {
-    for(unsigned int i=0; i<Vector3D::dimension; i++){
+    for(size_t i=0; i<Vector3D::dimension; i++){
         this->set_item(i, in.get_item(i));
     }
 }
